requirements.txt
apscheduler==2.1.1
d2to1==0.2.11
pbr==0.5.21
MySQL-python==1.2.3c1
SQLAlchemy==0.7.9
alembic==0.6.0
anyjson==0.3.3
iso8601==0.1.8
netaddr==0.7.5
argparse==1.2.1
eventlet==0.13.0
python-keystoneclient==0.4.1
python-ceilometerclient==1.0.6
python-novaclient==2.15.0
python-glanceclient==0.12.0
python-cinderclient==1.0.7
pecan==0.3.2
WSME==0.5b5
oslo.config==1.2.0
python-memcached==1.43
kombu==2.4.8
stevedore==0.11
